Generates route for places
Used `generate route` to make all the necessary files for places.
Have just text in the `places.hbs` and liked it to `my-application.hbs`
to see if the text would show up

Had to npm install `chalk` to allow `generate route` to work on this
template.

Still have to make a model to display places