Generates place model adds findAll
Generates the model for places and adds `findAll` to the `places.js` so
that all the places for that user are displayed.

Adds an each in the `places.hbs` so that each place in listed for the
user when they clock the temporary link.

Items still have not been worked on