Adds places/new route
Generate `places/new` route to be able to create. Changes the route
to have a function that includes the new route. Adds like to main
`places` page for add new place which link the the `places/new`.

Adds model for `newPlace.save` so that the user can input the name of a
place and have it save to the list.