Fixes multiple issues
First adds a component called `travel-list` that now houses the items
and the save and delete actions for that item. Had to be separated
because in the `place` there was an issue when trying to add items from
there

Second removes the RSVP.hash from `place` because it was bypassing the
model hook so that when you clicked on the title to take you to a new
page to display info for that `place` it would not load unless refreshed
commented out the actions related to items directly

Third moved the display of the items from `place` to `travel-list` to be
able to work on the items without interrupting the `place`

Had to change `place` to `model` in various places because place was
undefined.