Adds new route and messages
Added a `home-message` route so that any time someone clicks on home
it link to an explanation of the application. Added new fonts and
colors. Adds an iff statement in `place.hbs` to see if there are any
places. If there are no places the messages says to start by adding
a place. If there are places it shows a list of them.